chore: Add missing exports for internal utils
