+++ src/container-queries/interfaces.ts
@@ -2,22 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React from 'react';
-
-/**
- * Simplified version of ResizeObserverEntry
- */
-export interface ContainerQueryEntry {
-  /** Target element */
-  target: Element;
-  /** Element's content box width */
-  contentBoxWidth: number;
-  /** Element's content box height */
-  contentBoxHeight: number;
-  /** Element's border box width */
-  borderBoxWidth: number;
-  /** Element's border box height */
-  borderBoxHeight: number;
-}
+export { ContainerQueryEntry } from '../internal/container-queries/interfaces';
 
 /**
  * React reference or element callback
