+++ src/internal/index.ts
@@ -5,5 +5,9 @@ export { useComponentMetadata, COMPONENT_METADATA_KEY } from './base-component/c
 export { useComponentMetrics } from './base-component/component-metrics';
 export { initAwsUiVersions } from './base-component/init-awsui-versions';
 export { Metrics } from './base-component/metrics/metrics';
+export { useResizeObserver } from './container-queries/use-resize-observer';
+export { createSingletonHandler, createSingletonState, UseSingleton } from './singleton-handler';
+export { useStableCallback } from './stable-callback';
+export { isMotionDisabled, useCurrentMode, useDensityMode, useReducedMotion } from './visual-mode';
 export { isDevelopment } from './is-development';
 export { warnOnce } from './logging';
